[virt] kvm/vmx: Allow the guest to run with dirty debug registers
Message-id: <1406040016-3289-163-git-send-email-pbonzini@redhat.com>
Patchwork-id: 86241
O-Subject: [RHEL7 PATCH v2 162/221] KVM: vmx: Allow the guest to run with dirty debug registers
Bugzilla: 1068627 1116936
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When not running in guest-debug mode (i.e. the guest controls the debug
registers, having to take an exit for each DR access is a waste of time.
If the guest gets into a state where each context switch causes DR to be
saved and restored, this can take away as much as 40% of the execution
time from the guest.

If the guest is running with vcpu->arch.db == vcpu->arch.eff_db, we
can let it write freely to the debug registers and reload them on the
next exit.  We still need to exit on the first access, so that the
KVM_DEBUGREG_WONT_EXIT flag is set in switch_db_regs; after that, further
accesses to the debug registers will not cause a vmexit.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 81908bf44340eb5ebc9969f67e6c8be0c92f2857)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>