[virt] kvm/x86: handle missing MPX in nested virtualization
Message-id: <1406040016-3289-168-git-send-email-pbonzini@redhat.com>
Patchwork-id: 86246
O-Subject: [RHEL7 PATCH v2 167/221] KVM: x86: handle missing MPX in nested virtualization
Bugzilla: 1116936
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

When doing nested virtualization, we may be able to read BNDCFGS but
still not be allowed to write to GUEST_BNDCFGS in the VMCS.  Guard
writes to the field with vmx_mpx_supported(), and similarly hide the
MSR from userspace if the processor does not support the field.

We could work around this with the generic MSR save/load machinery,
but there is only a limited number of MSR save/load slots and it is
not really worthwhile to waste one for a scenario that should not
happen except in the nested virtualization case.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 93c4adc7afedf9b0ec190066d45b6d67db5270da)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>