[virt] virtio-ccw: virtio-ccw adapter interrupt support
Message-id: <1406112611-9862-14-git-send-email-brueckner@redhat.com>
Patchwork-id: 86324
O-Subject: [RHEL7.1 PATCH 13/23] [s390] virtio-ccw: virtio-ccw adapter interrupt support.
Bugzilla: 1088554
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
commit 96b14536d935848cffd904f583f67c66169002d8
Author: Cornelia Huck <cornelia.huck@de.ibm.com>
Commit: Christian Borntraeger <borntraeger@de.ibm.com>

    virtio-ccw: virtio-ccw adapter interrupt support.

    Implement the new CCW_CMD_SET_IND_ADAPTER command and try to enable
    adapter interrupts for every device on the first startup. If the host
    does not support adapter interrupts, fall back to normal I/O interrupts.

    virtio-ccw adapter interrupts use the same isc as normal I/O subchannels
    and share a summary indicator for all devices sharing the same indicator
    area.

    Indicator bits for the individual virtqueues may be contained in the same
    indicator area for different devices.

    Signed-off-by: Cornelia Huck <cornelia.huck@de.ibm.com>
    Signed-off-by: Christian Borntraeger <borntraeger@de.ibm.com>

Bugzilla
--------
BZ 1088554
https://bugzilla.redhat.com/show_bug.cgi?id=1088554

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version v3.15-rc1
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=96b14536d935848cffd904f583f67c66169002d8

kABI
----
Introducing a new IRQ sub description to account IRQ statistics causes a
kABI breakage for the irq_stat symbol.  Instead of adding the new IRQ stat
counter in the middle of the enum, it is appended to the end and excluded
for kABI checking.  A follow-up commit ensures that the order in which the
counters are reported in /proc/interrupts is not changed (compared to the
upstream kernel).

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=7725759

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Jarod Wilson <jarod@redhat.com>