[fs] nfs: nfs_page should take a ref on the head req
Message-id: <1406225773-16395-6-git-send-email-steved@redhat.com>
Patchwork-id: 86639
O-Subject: [RHEL7.1 PATCH 05/10] nfs: nfs_page should take a ref on the head req
Bugzilla: 1120728
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Weston Andros Adamson <dros@primarydata.com>

nfs_pages that aren't the the head of a group must take a reference on the
head as long as ->wb_head is set to it. This stops the head from hitting
a refcount of 0 while there is still an active nfs_page for the page group.

This avoids kref warnings in the writeback code when the page group head
is found and referenced.

Signed-off-by: Weston Andros Adamson <dros@primarydata.com>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream Id: 85710a837c2026aae80b7c64187edf1f10027b0b
Bugzilla: 1120728

Signed-off-by: Jarod Wilson <jarod@redhat.com>