[drm] qxl: return IRQ_NONE if it was not our irq
Message-id: <1406775111-19208-1-git-send-email-jasowang@redhat.com>
Patchwork-id: 86855
O-Subject: [RHEL7.1 PATCH] drm/qxl: return IRQ_NONE if it was not our irq
Bugzilla: 1095099
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

Bugzilla: 1095099
Brew Build: https://brewweb.devel.redhat.com/taskinfo?taskID=7764225
Test status: Tested by QE

Return IRQ_NONE if it was not our irq. This is necessary for the case
when qxl is sharing irq line with a device A in a crash kernel. If qxl
is initialized before A and A's irq was raised during this gap,
returning IRQ_HANDLED in this case will cause this irq to be raised
again after EOI since kernel think it was handled but in fact it was
not.

Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: stable@vger.kernel.org
Signed-off-by: Jason Wang <jasowang@redhat.com>
Signed-off-by: Dave Airlie <airlied@redhat.com>
(cherry picked from commit fbb60fe35ad579b511de8604b06a30b43846473b)
Signed-off-by: Jarod Wilson <jarod@redhat.com>