[fs] gfs2: Remove gfs2_quota_change_host structure
Message-id: <1408132603-52384-11-git-send-email-adas@redhat.com>
Patchwork-id: 87818
O-Subject: [RHEL7.1 PATCH 10/17] GFS2: Remove gfs2_quota_change_host structure
Bugzilla: 1026406
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

There is only one place this is used, when reading in the quota
changes at mount time. It is not really required and much
simpler to just convert the fields from the on-disk structure
as required.

There should be no functional change as a result of this patch.

Resolves: rhbz#1026406
Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>