[hwrng] create filler thread
Message-id: <d173349d2a97f11b8ace246bc7b4bd2f9f430e7c.1408347166.git.amit.shah@redhat.com>
Patchwork-id: 87839
O-Subject: [RHEL7.1 kernel PATCH 2/6] hwrng: create filler thread
Bugzilla: 1129195
RH-Acked-by: Amos Kong <akong@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Herbert Xu <herbert.xu@redhat.com>

From: Torsten Duwe <duwe@lst.de>

This can be viewed as the in-kernel equivalent of hwrngd;
like FUSE it is a good thing to have a mechanism in user land,
but for some reasons (simplicity, secrecy, integrity, speed)
it may be better to have it in kernel space.

This patch creates a thread once a hwrng registers, and uses
the previously established add_hwgenerator_randomness() to feed
its data to the input pool as long as needed. A derating factor
is used to bias the entropy estimation and to disable this
mechanism entirely when set to zero.

Signed-off-by: Torsten Duwe <duwe@suse.de>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Acked-by: H. Peter Anvin <hpa@zytor.com>
(cherry picked from commit be4000bc4644d027c519b6361f5ae3bbfc52c347)
Signed-off-by: Amit Shah <amit.shah@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>

Conflicts:
	drivers/char/hw_random/core.c
 Trivial contextual conflict -- also happened upstream in the -next tree
 as two series were modifying the same location.