[x86] perf: Report TSX transaction abort cost as weight
Message-id: <1408462094-14194-3-git-send-email-jolsa@redhat.com>
Patchwork-id: 87990
O-Subject: [PATCH RHEL7.1 BZ1131394 002/325] perf/x86: Report TSX transaction abort cost as weight
Bugzilla: 1131394
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1131394
https://bugzilla.redhat.com/show_bug.cgi?id=1131394

upstream
========
commit 748e86aa90edfddfa6016f1cf383ff5bc6aada91
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Sep 5 20:37:39 2013 -0700

description
===========
Use the existing weight reporting facility to report the transaction
abort cost, that is the number of cycles wasted in aborts.
Haswell reports this in the PEBS record.

This was in fact the original user for weight.

This is a very useful sort key to concentrate on the most
costly aborts and a good metric for TSX tuning.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>