[tools] perf/trace: Show path associated with fd in live sessions
Message-id: <1408462094-14194-142-git-send-email-jolsa@redhat.com>
Patchwork-id: 88108
O-Subject: [PATCH RHEL7.1 BZ1131394 141/325] perf trace: Show path associated with fd in live sessions
Bugzilla: 1131394
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1131394
https://bugzilla.redhat.com/show_bug.cgi?id=1131394

upstream
========
commit 75b757ca90469e990e6901f4a9497fe4161f7f5a
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Sep 24 11:04:32 2013 -0300

description
===========
For live sessions we can just access /proc to map an fd to its path, on
a best effort way, i.e. sometimes the fd will have gone away when we try
to do the mapping, as it is done in a lazy way, only when a reference to
such fd is made then the path will be looked up in /proc.

This is disabled when processing perf.data files, where we will have to
have a way to get getname events, be it via an on-the-fly 'perf probe'
event or after a vfs_getname tracepoint is added to the kernel.

A first step will be to synthesize such event for the use cases where
the threads in the monitored workload exist already.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>