[tools] perf/symbols: Fix a mmap and munmap mismatched bug
Message-id: <1408462094-14194-193-git-send-email-jolsa@redhat.com>
Patchwork-id: 88177
O-Subject: [PATCH RHEL7.1 BZ1131394 192/325] perf symbols: Fix a mmap and munmap mismatched bug
Bugzilla: 1131394
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1131394
https://bugzilla.redhat.com/show_bug.cgi?id=1131394

upstream
========
commit 784f3390f9bd900adfb3b0373615e105a0d9749a
Author: Chenggang Qin <chenggang.qcg@taobao.com>
Date: Fri Oct 11 08:27:57 2013 +0800

description
===========
In function filename__read_debuglink(), while the ELF file is opend and
mmapped in elf_begin(), but if this file is considered to not be usable
during the following code, we will goto the close(fd) directly. The
elf_end() is skipped.  So, the mmaped ELF file cannot be munmapped. The
mmapped areas exist during the life of perf.

This is a memory leak.  This patch fixed this bug.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>